+++ woodwork/statistics_utils.py
@@ -463,17 +463,17 @@ def _get_box_plot_info_for_column(
             Defaults to True.
 
     Note:
-        The minimum quantiles necessary for outlier detection using the IQR method are the
-        first quartile (0.25) and third quartile (0.75). If these keys are missing from the quantiles dictionary,
-        the following quantiles will be calculated: {0.0, 0.25, 0.5, 0.75, 1.0}, which correspond to
-        {min, first quantile, median, third quantile, max}.
+        The minimum quantiles necessary for building a box plot using the IQR method are the
+        minimum value (0.0 in the quantiles dict), first quartile (0.25), third quartile (0.75), and maximum value (1.0).
+        If no quantiles are provided, the following quantiles will be calculated:
+        {0.0, 0.25, 0.5, 0.75, 1.0}, which correspond to {min, first quantile, median, third quantile, max}.
 
     Returns:
         (dict[str -> float,list[number]]): Returns a dictionary containing box plot information for the Series.
             The following elements will be found in the dictionary:
 
-            - low_bound (float): the lower bound below which outliers lay - to be used as a whisker
-            - high_bound (float): the high bound above which outliers lay - to be used as a whisker
+            - low_bound (float): the lowest data point in the dataset excluding any outliers - to be used as a whisker
+            - high_bound (float): the highest point in the dataset excluding any outliers - to be used as a whisker
             - quantiles (list[float]): the quantiles used to determine the bounds.
                 If quantiles were passed in, will contain all quantiles passed in. Otherwise, contains the five
                 quantiles {0.0, 0.25, 0.5, 0.75, 1.0}.
@@ -526,46 +526,36 @@ def _get_box_plot_info_for_column(
     # calculate the outlier bounds using IQR
     if quantiles is None:
         quantiles = series.quantile([0.0, 0.25, 0.5, 0.75, 1.0]).to_dict()
-    elif 0.25 not in quantiles or 0.75 not in quantiles:
+    elif len(set(quantiles.keys()) & {0.0, 0.25, 0.75, 1.0}) != 4:
         raise ValueError(
-            "Input quantiles do not contain the minimum necessary quantiles for outlier calculation: "
-            "0.25 (the first quartile) and 0.75 (the third quartile)."
+            "Input quantiles do not contain the minimum necessary quantiles for box plot calculation: "
+            "0.0 (the minimum value), 0.25 (the first quartile), 0.75 (the third quartile), and 1.0 (the maximum value)."
         )
-
+    min_value = quantiles[0.0]
     q1 = quantiles[0.25]
     q3 = quantiles[0.75]
+    max_value = quantiles[1.0]
 
+    # Box plot bounds calculation - the bounds should never be beyond the min and max values
     iqr = q3 - q1
-
-    low_bound = q1 - (iqr * 1.5)
-    high_bound = q3 + (iqr * 1.5)
+    low_bound = max(q1 - (iqr * 1.5), min_value)
+    high_bound = min(q3 + (iqr * 1.5), max_value)
 
     if include_indices_and_values:
         # identify outliers in the series
-        min = quantiles.get(0.0)
-        max = quantiles.get(1.0)
-        if (
-            min is not None
-            and max is not None
-            and low_bound <= min
-            and high_bound >= max
-        ):
-            outliers_dict = {
-                "low_values": [],
-                "high_values": [],
-                "low_indices": [],
-                "high_indices": [],
-            }
-        else:
-            low_series = series[series < low_bound]
-            high_series = series[series > high_bound]
+        low_series = (
+            series[series < low_bound] if low_bound > min_value else pd.Series()
+        )
+        high_series = (
+            series[series > high_bound] if high_bound < max_value else pd.Series()
+        )
 
-            outliers_dict = {
-                "low_values": low_series.tolist(),
-                "high_values": high_series.tolist(),
-                "low_indices": low_series.index.tolist(),
-                "high_indices": high_series.index.tolist(),
-            }
+        outliers_dict = {
+            "low_values": low_series.tolist(),
+            "high_values": high_series.tolist(),
+            "low_indices": low_series.index.tolist(),
+            "high_indices": high_series.index.tolist(),
+        }
 
     return {
         "low_bound": low_bound,