+++ docs/source/release_notes.rst
@@ -11,6 +11,7 @@ Future Release
         * Fixed issue with tuples being incorrectly inferred as EmailAddress (:pr:`1253`)
     * Changes
         * Prevent setting index that contains null values (:pr:`1239`)
+        * Allow tuple NaN LatLong values (:pr:`1255`)
         * Update ipython to 7.31.1 (:pr:`1258`)
         * Temporarily restrict pandas and koalas max versions (:pr:`1261`)
         * Update to drop Python 3.7 support and add support for pandas version 1.4.0 (:pr:`1264`)